Change the way sizeof is used (and also change the type of tile attrs)
By using sizeof(*array) I remove the responsibility of having to go over
every allocation of the array when changing the type of its elements.

+++ include/board.h
@@ -39,7 +39,7 @@ inline ALWAYS_INLINE bool get_tile_attr(const Game* game, Coords coords, int att
 
 inline ALWAYS_INLINE void set_tile_attr(Game* game, Coords coords, int attr, bool value) {
   assert(is_tile_in_bounds(game, coords));
-  int* ptr = &game->tile_attributes[coords.x + game->board_width * coords.y];
+  short* ptr = &game->tile_attributes[coords.x + game->board_width * coords.y];
   *ptr = change_bit(*ptr, 1 << attr, value);
 }
